+++ src/main/java/com/mystdev/recicropal/client/ClientEvents.java
@@ -19,7 +19,7 @@ import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;
 
 @Mod.EventBusSubscriber(modid = Recicropal.MOD_ID, bus = Mod.EventBusSubscriber.Bus.FORGE, value = Dist.CLIENT)
-public class Events {
+public class ClientEvents {
 
     @SubscribeEvent
     public static void renderTooltip(RenderTooltipEvent.GatherComponents event) {